test(web-client): split unit and browser-backed vitest projects
