+++ apps/web-client/tsconfig.json
@@ -2,6 +2,7 @@
 	"extends": "./.svelte-kit/tsconfig.json",
 	"compilerOptions": {
 		"types": ["svelte", "@sveltejs/kit", "@vitest/browser/matchers"],
-		"skipLibCheck": true
+		"skipLibCheck": true,
+		"verbatimModuleSyntax": true
 	}
 }
