+++ common/config/rush/repo-state.json
@@ -1,5 +1,5 @@
 // DO NOT MODIFY THIS FILE MANUALLY BUT DO COMMIT IT. It is generated and used by Rush.
 {
-  "pnpmShrinkwrapHash": "98b3623142e58eeab344ddb0d7bae84d0e2d6d83",
+  "pnpmShrinkwrapHash": "7f5b07b74e68deb5dc38026d919dea42501ac8cd",
   "preferredVersionsHash": "bf21a9e8fbc5a3846fb05b4fa0859e0917b2202f"
 }
